option to leave the fits file

+++ pizza_cutter/coadd_sim_slicer/slicer.py
@@ -60,7 +60,8 @@ def make_meds_pizza_slices(
         seg_path=None, seg_ext=None,
         psf,
         fpack_pars=None,
-        seed):
+        seed,
+        remove_fits_file=True):
     """Build a MEDS pizza slices file.
 
     Parameters
@@ -103,6 +104,8 @@ def make_meds_pizza_slices(
         A dictionary of fpack header keywords for compression.
     seed : int
         The random seed used to make the noise field.
+    remove_fits_file : bool, optional
+        If `True`, remove the FITS file after fpacking.
     """
 
     metadata = _build_metadata(config)
@@ -150,7 +153,8 @@ def make_meds_pizza_slices(
     except Exception:
         pass
     else:
-        os.remove(meds_path)
+        if remove_fits_file:
+            os.remove(meds_path)
 
     # validate the fpacked file
     print('validating:', flush=True)
